+++ src/Service/Cryptonator.php
@@ -16,11 +16,15 @@ namespace Exchanger\Service;
 use Exchanger\Contract\ExchangeRateQuery;
 use Exchanger\Contract\HistoricalExchangeRateQuery;
 use Exchanger\Exception\Exception;
-use Exchanger\ExchangeRate;
 use Exchanger\StringUtil;
 use Exchanger\Contract\ExchangeRate as ExchangeRateContract;
 
-final class Cryptonator extends Service
+/**
+ * The Cryptonator service.
+ *
+ * @author Danny Weeks
+ */
+final class Cryptonator extends HttpService
 {
     const LATEST_URL = 'https://api.cryptonator.com/api/ticker/%s-%s';
 
@@ -55,7 +59,7 @@ final class Cryptonator extends Service
 
         $date = (new \DateTime())->setTimestamp($data['timestamp']);
 
-        return new ExchangeRate((float) ($data['ticker']['price']), __CLASS__, $date);
+        return $this->createRate($currencyPair, (float) ($data['ticker']['price']), $date);
     }
 
     /**
@@ -67,9 +71,11 @@ final class Cryptonator extends Service
      */
     public function supportQuery(ExchangeRateQuery $exchangeQuery): bool
     {
+        $currencyPair = $exchangeQuery->getCurrencyPair();
+
         return !$exchangeQuery instanceof HistoricalExchangeRateQuery
-            && in_array($exchangeQuery->getCurrencyPair()->getBaseCurrency(), $this->getSupportedCodes())
-            && in_array($exchangeQuery->getCurrencyPair()->getQuoteCurrency(), $this->getSupportedCodes());
+            && in_array($currencyPair->getBaseCurrency(), $this->getSupportedCodes())
+            && in_array($currencyPair->getQuoteCurrency(), $this->getSupportedCodes());
     }
 
     /**
@@ -82,5 +88,13 @@ final class Cryptonator extends Service
     private function getSupportedCodes(): array
     {
         return require __DIR__.'/resources/cryptonator-codes.php';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName(): string
+    {
+        return 'cryptonator';
     }
 }