+++ src/Service/CurrencyConverterApi.php
@@ -27,16 +27,12 @@ use Exchanger\Contract\ExchangeRate as ExchangeRateContract;
 /**
  * CurrencyConverterApi service.
  *
- * @see https://github.com/florianv/laravel-swap/tree/master/doc#creating-a-service
- *
- * Docs for free plan:
- * @see https://free.currencyconverterapi.com/
- *
- * Docs for paid/enterprise plans:
- * @see https://www.currencyconverterapi.com/docs
+ * @author Aliaksandr lptn
  */
-final class CurrencyConverterApi extends HistoricalService
+final class CurrencyConverterApi extends HttpService
 {
+    use SupportsHistoricalQueries;
+
     const FREE_LATEST_URL = 'https://free.currencyconverterapi.com/api/v6/convert?q=%s';
 
     const ENTERPRISE_LATEST_URL = 'https://api.currencyconverterapi.com/api/v6/convert?q=%s&apiKey=%s';
@@ -152,6 +148,7 @@ final class CurrencyConverterApi extends HistoricalService
         $currencyPair = $exchangeRateQuery->getCurrencyPair();
 
         $response = $this->getResponse($url);
+
         if (200 !== $response->getStatusCode()) {
             throw new Exception("Unexpected response status {$response->getReasonPhrase()}, $url");
         }
@@ -159,6 +156,7 @@ final class CurrencyConverterApi extends HistoricalService
         $responsePayload = StringUtil::jsonToArray($response->getBody()->__toString());
 
         $keyAsCurrencyPair = $this->stringifyCurrencyPair($currencyPair);
+
         if (empty($responsePayload['results'][$keyAsCurrencyPair]['val'])) {
             throw new Exception("Unexpected response body {$response->getReasonPhrase()}");
         }
@@ -173,14 +171,14 @@ final class CurrencyConverterApi extends HistoricalService
 
         if ($exchangeRateQuery instanceof HistoricalExchangeRateQuery) {
             $dateStringified = $responsePayload['date'];
-            $date = new DateTime($dateStringified);
+            $date = new \DateTime($dateStringified);
             $rate = $responsePayload['results'][$keyAsCurrencyPair]['val'][$dateStringified];
         } else {
-            $date = new DateTime('now');
+            $date = new \DateTime('now');
             $rate = $responsePayload['results'][$keyAsCurrencyPair]['val'];
         }
 
-        return new ExchangeRate((float) $rate, __CLASS__, $date);
+        return $this->createRate($currencyPair, (float) $rate, $date);
     }
 
     /**
@@ -196,16 +194,16 @@ final class CurrencyConverterApi extends HistoricalService
     /**
      * Gets the earliest available date for the historical query.
      *
-     * @return DateTime
+     * @return \DateTime
      */
-    private function getEarliestAvailableDateForHistoricalQuery(): DateTime
+    private function getEarliestAvailableDateForHistoricalQuery(): \DateTime
     {
         if ($this->isEnterprise()) {
-            return (new DateTime())->setTimestamp(0);
+            return (new \DateTime())->setTimestamp(0);
         }
 
         // Historical rates for free plan is available only for the past year.
-        return new DateTime('-1 year 00:00');
+        return new \DateTime('-1 year 00:00');
     }
 
     /**
@@ -226,10 +224,20 @@ final class CurrencyConverterApi extends HistoricalService
      *
      * @param DateTimeInterface $dateTime
      *
-     * @return DateTime
+     * @return \DateTime
      */
-    private function getAdoptedDateTime(DateTimeInterface $dateTime): DateTime
+    private function getAdoptedDateTime(DateTimeInterface $dateTime): \DateTime
     {
-        return (new DateTime())->setTimestamp($dateTime->getTimestamp())->setTimezone(new DateTimeZone('Asia/Manila'));
+        return (new \DateTime())
+            ->setTimestamp($dateTime->getTimestamp())
+            ->setTimezone(new DateTimeZone('Asia/Manila'));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName(): string
+    {
+        return 'currency_converter';
     }
 }