+++ tests/Tests/Service/CentralBankOfCzechRepublicTest.php
@@ -78,62 +78,84 @@ class CentralBankOfCzechRepublicTest extends ServiceTestCase
     /**
      * @test
      */
-    public function itFetchesAEurRate()
+    public function it_fetches_eur_rate()
     {
         $service = $this->createService();
-        $rate = $service->getExchangeRate(new ExchangeRateQuery(CurrencyPair::createFromString('EUR/CZK')));
+        $pair = CurrencyPair::createFromString('EUR/CZK');
+        $rate = $service->getExchangeRate(new ExchangeRateQuery($pair));
 
         $this->assertSame(27.035, $rate->getValue());
         $this->assertEquals(new \DateTime('2016-04-05'), $rate->getDate());
-        $this->assertEquals(CentralBankOfCzechRepublic::class, $rate->getProvider());
+        $this->assertEquals('central_bank_of_czech_republic', $rate->getProviderName());
+        $this->assertSame($pair, $rate->getCurrencyPair());
     }
 
     /**
      * @test
      */
-    public function itFetchesAPhpRate()
+    public function it_fetches_php_rate()
     {
-        $rate = $this->createService()->getExchangeRate(new ExchangeRateQuery(CurrencyPair::createFromString('PHP/CZK')));
+        $pair = CurrencyPair::createFromString('PHP/CZK');
+        $rate = $this->createService()->getExchangeRate(new ExchangeRateQuery($pair));
+
         $this->assertSame(0.51384, $rate->getValue());
-        $this->assertEquals(CentralBankOfCzechRepublic::class, $rate->getProvider());
+        $this->assertEquals('central_bank_of_czech_republic', $rate->getProviderName());
+        $this->assertSame($pair, $rate->getCurrencyPair());
     }
 
     /**
      * @test
      */
-    public function itFetchesAIdrRate()
+    public function it_fetches_idr_rate()
     {
-        $rate = $this->createService()->getExchangeRate(new ExchangeRateQuery(CurrencyPair::createFromString('IDR/CZK')));
+        $pair = CurrencyPair::createFromString('IDR/CZK');
+        $rate = $this->createService()->getExchangeRate(new ExchangeRateQuery($pair));
+
         $this->assertSame(0.001798, $rate->getValue());
-        $this->assertEquals(CentralBankOfCzechRepublic::class, $rate->getProvider());
+        $this->assertEquals('central_bank_of_czech_republic', $rate->getProviderName());
+        $this->assertSame($pair, $rate->getCurrencyPair());
     }
 
     /**
      * @test
      */
-    public function itFetchesAHistoricalFrfRate()
+    public function it_fetches_historical_frf_rate()
     {
         $requestedDate = new \DateTime('2000-04-23');
         $service = $this->createServiceForHistoricalRates();
-        $rate = $service->getExchangeRate(new HistoricalExchangeRateQuery(CurrencyPair::createFromString('FRF/CZK'), $requestedDate));
+        $pair = CurrencyPair::createFromString('FRF/CZK');
+        $rate = $service->getExchangeRate(new HistoricalExchangeRateQuery($pair, $requestedDate));
 
         $this->assertEquals(5.529, $rate->getValue());
         $this->assertEquals(new \DateTime('2000-04-21'), $rate->getDate());
-        $this->assertEquals(CentralBankOfCzechRepublic::class, $rate->getProvider());
+        $this->assertEquals('central_bank_of_czech_republic', $rate->getProviderName());
+        $this->assertSame($pair, $rate->getCurrencyPair());
     }
 
     /**
      * @test
      */
-    public function itFetchesAHistoricalEurRate()
+    public function it_fetches_historical_eur_rate()
     {
         $requestedDate = new \DateTime('2000-04-23');
         $service = $this->createServiceForHistoricalRates();
-        $rate = $service->getExchangeRate(new HistoricalExchangeRateQuery(CurrencyPair::createFromString('EUR/CZK'), $requestedDate));
+        $pair = CurrencyPair::createFromString('EUR/CZK');
+        $rate = $service->getExchangeRate(new HistoricalExchangeRateQuery($pair, $requestedDate));
 
         $this->assertEquals(36.27, $rate->getValue());
         $this->assertEquals(new \DateTime('2000-04-21'), $rate->getDate());
-        $this->assertEquals(CentralBankOfCzechRepublic::class, $rate->getProvider());
+        $this->assertEquals('central_bank_of_czech_republic', $rate->getProviderName());
+        $this->assertSame($pair, $rate->getCurrencyPair());
+    }
+
+    /**
+     * @test
+     */
+    public function it_has_a_name()
+    {
+        $service = new CentralBankOfCzechRepublic($this->createMock('Http\Client\HttpClient'));
+
+        $this->assertSame('central_bank_of_czech_republic', $service->getName());
     }
 
     /**